+++ src/Components/Account/logout.jsx
@@ -2,8 +2,8 @@ import * as React from "react";
 import List from "@mui/material/List";
 import Divider from "@mui/material/Divider";
 import LogoutIcon from "@mui/icons-material/Logout";
-import { Title } from "../../Styled/Styles";
-import { ListItems } from "../../Styled/Material";
+import { Title } from "../../Styled/ComponentUI/Styles";
+import { ListItems } from "../../Styled/MUI/AccountStyle.js";
 
 export default function Logout() {
   return (
