+++ src/Components/Footer/Footer.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { Box, Container, Row, Column, FooterLink, Heading } from "./FooterStyles";
+import { Box, Container, Row, Column, FooterLink, Heading } from "../../Styled/ComponentUI/FooterStyles";
 import FacebookIcon from "@mui/icons-material/Facebook";
 import InstagramIcon from "@mui/icons-material/Instagram";
 import TwitterIcon from "@mui/icons-material/Twitter";
