+++ src/Components/GoogleLogin/GoogleLogin.jsx
@@ -1,12 +1,12 @@
 import React from "react";
 import { useGoogleLogin } from "@react-oauth/google";
-import { useDispatch } from "react-redux";
+// import { useDispatch } from "react-redux";
 import googleLogo from "../../assets/images/googleLogo.svg";
-import { Buttonx, Dividers } from "../../Pages/Register/AuthStyles";
-import { loginWithGoogle } from "../../Redux/Actions/authAction";
+import { Buttonx, Dividers } from "../../Styled/MUI/AuthStyles";
+// import { loginWithGoogle } from "../../Redux/Actions/authAction";
 
 function GoogleLogin({ label }) {
-  const dispatch = useDispatch();
+  // const dispatch = useDispatch();
 
   const googleLogin = useGoogleLogin({
     onSuccess: async (res) => {
