+++ src/Components/Header/TopPanel.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { TopPanel } from "../../Styled/Material";
+import { TopPanel } from "../../Styled/MUI/HeaderStyle";
 // import sample from "../../assets/images/terbang.gif";
 
 function TopPanels() {
