+++ src/Pages/Account/Profile/Buyers.jsx
@@ -12,12 +12,14 @@ import ReceiptIcon from "@mui/icons-material/Receipt";
 import Divider from "@mui/material/Divider";
 import Logout from "../../../Components/Account/logout";
 import Footer from "../../../Components/Footer/Footer";
-import { BoxHome, BoxMenu, BoxPage, BoxTitle, Buttons, TextFields } from "../../../Styled/Material";
-import { CardMenu, CardBox, CardPage } from "../../../Styled/Material";
-import { Links, ListItems, ListPage } from "../../../Styled/Material";
-import { Heading, Title } from "../../../Styled/Styles";
+import { BoxHome, BoxMenu, BoxPage, BoxTitle, Buttons, TextFields } from "../../../Styled/MUI/AccountStyle.js";
+import { CardMenu, CardBox, CardPage } from "../../../Styled/MUI/AccountStyle.js";
+import { Links, ListItems, ListPage } from "../../../Styled/MUI/AccountStyle.js";
+import { Heading, Title } from "../../../Styled/ComponentUI/Styles";
+import TopPanels from "../../../Components/Header/TopPanel";
 
 function Buyers() {
+  const [gender, setGender] = useState("Gender");
   const titles = [
     {
       value: "Male",
@@ -29,13 +31,12 @@ function Buyers() {
     },
   ];
 
-  const [gender, setGender] = useState("Gender");
-
   const handleChange = (event) => {
     setGender(event.target.value);
   };
   return (
     <>
+      <TopPanels />
       <NavBar />
       <BoxHome>
         <BoxPage>
@@ -112,17 +113,19 @@ function Buyers() {
                 </Grid>
                 <CardBox variant="outlined">
                   <CardContent>
-                    <TextFields id="outlined-basic" variant="outlined" size="medium" select label="Select" value={gender} onChange={handleChange} helperText="Please select your gender">
+                    <TextFields id="outlined-basic" label="First Name" variant="outlined" size="small" helperText="Please enter your first name" />
+                    <TextFields id="outlined-basic" label="Last Name" variant="outlined" size="small" helperText="Please enter your last name" />
+                    <TextFields id="outlined-basic" variant="outlined" size="small" select label="Select" value={gender} onChange={handleChange} helperText="Please select your gender">
                       {titles.map((option) => (
                         <MenuItem key={option.value} value={option.value}>
                           {option.label}
                         </MenuItem>
                       ))}
                     </TextFields>
-                    <TextFields id="outlined-basic" label="First Name" variant="outlined" size="medium" helperText="Please enter your first name" />
-                    <TextFields id="outlined-basic" label="Last Name" variant="outlined" size="medium" helperText="Please enter your last name" />
-                    <TextFields id="outlined-basic" label="ID Number" variant="outlined" size="medium" helperText="Please enter your ID card number" />
-                    <Buttons id="outlined-basic" label="button" variant="contained" size="medium">
+                    <TextFields id="outlined-basic" label="Phone Number" variant="outlined" size="small" helperText="Please enter your Phone number" />
+                    <TextFields id="outlined-basic" label="Address" variant="outlined" size="small" helperText="Please enter your address" />
+                    <TextFields id="outlined-basic" label="Nationality" variant="outlined" size="small" helperText="Please enter your Nationality" />
+                    <Buttons id="outlined-basic" label="button" variant="contained" size="small">
                       Save
                     </Buttons>
                   </CardContent>
